libfslcrypto, simple_crypto, simple_proto, fsl_sec: Fix KW Sev8&9 issues
This patch fixes all-around the place the issues identified by KW
with Severity 8&9:
- Comparison of unsigned value against 0 is always true
- Value is Never Used after Assignment
- Loss of Precision
- Loss of Precision during function call

Reviewed-on: http://git.am.freescale.net:8181/19441
Tested-by: Vakul Garg <vakul@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22589

+++ apps/simple_crypto/include/simple_crypto.h
@@ -103,8 +103,8 @@ void init_rtv_snow_f8_f9(struct test_param *crypto_info);
 /* prepare test buffers, fqs, fds routines */
 int prepare_test_frames(struct test_param *crypto_info);
 void set_buf_size(struct test_param *crypto_info);
-void *setup_preheader(uint32_t shared_desc_len, uint32_t pool_id,
-		      uint32_t pool_buf_size, uint8_t absolute,
+void *setup_preheader(uint32_t shared_desc_len, uint8_t pool_id,
+		      uint8_t pool_buf_size, uint8_t absolute,
 		      uint8_t add_buf);
 static void *setup_init_descriptor(bool mode, struct test_param crypto_info);
 void *setup_sec_descriptor(bool mode, void *params);
